requirements.txt
# LLM Local
ollama==0.1.6

# Embeddings Local
sentence-transformers>=2.6.0
torch>=2.0.0

# Vector Store Local
chromadb>=0.5.0

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.0
numpy<2.0